// src/app/api-client/model/licensedAsset.ts
/**
 * Horde Model Reference API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LicenseAssignment } from './licenseAssignment';
import { LicensedAssetKind } from './licensedAssetKind';
import { LicensingRecordMetadata } from './licensingRecordMetadata';

/**
 * Represents a licensed non-model asset managed directly by PRIMARY editors.
 */
export interface LicensedAsset {
  asset_kind: LicensedAssetKind;
  asset_identifier: string;
  display_name: string;
  source_url?: string | null;
  version?: string | null;
  locations?: Array<string>;
  related_assets?: Array<string>;
  licensing: LicenseAssignment;
  notes?: string | null;
  metadata?: LicensingRecordMetadata;
}
export namespace LicensedAsset {}