// src/app/api-client/model/licensedAssetView.ts
/**
 * Horde Model Reference API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ModelLicensing } from './modelLicensing';
import { LicensingRecordMetadata } from './licensingRecordMetadata';
import { MODEL_REFERENCE_CATEGORY } from './mODELREFERENCECATEGORY';

/**
 * Represents a unified public view of a model or non-model licensed asset.
 */
export interface LicensedAssetView {
  asset_kind: string;
  asset_identifier: string;
  display_name: string;
  category?: MODEL_REFERENCE_CATEGORY | null;
  source_url?: string | null;
  version?: string | null;
  locations?: Array<string>;
  related_assets?: Array<string>;
  licensing: ModelLicensing;
  definition_urls?: { [key: string]: string };
  notes?: string | null;
  metadata?: LicensingRecordMetadata | null;
}
export namespace LicensedAssetView {}